Support for multiple eureka endpoints
This PR introduces failover support for the Eureka client.
Multiple endpoints can be specified in the `serviceUrl` array.
The endpoints will be tried one after another until it succeeds
in registering. The eureka endpoints should be in the same zone.
It will work even if they're not but behavior can be odd.

In contrast to how the client behaved before, it will keep
reconnecting to the servers it has (and register will not
issue the callback unless an endpoint explicitly gives an error, e.g.
500 Internal Server Error or similar).